fix(Security): 错误更新
