+++ admin4j-dependencies/pom.xml
@@ -155,7 +155,7 @@
             <dependency>
                 <groupId>com.admin4j.framework</groupId>
                 <artifactId>security-spring-boot-starter</artifactId>
-                <version>${admin4j.version}</version>
+                <version>0.9.2</version>
             </dependency>
             <dependency>
                 <groupId>com.admin4j.framework</groupId>
