+++ security-spring-boot-starter/pom.xml
@@ -10,7 +10,7 @@
     <groupId>com.admin4j.framework</groupId>
     <artifactId>security-spring-boot-starter</artifactId>
     <packaging>jar</packaging>
-    <!--    <version>0.9.1-SNAPSHOT</version>-->
+    <version>0.9.2</version>
 
     <name>security-spring-boot-starter</name>
